requirements.txt
Flask==2.3.2
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
Pillow==9.5.0
gunicorn==20.1.0
python-dotenv==1.0.0
requests==2.31.0